5.53/LayoutTests/http/tests/cache/disk-cache/manx/disk-cache-304-status-code.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var tests =
[
 { responseHeaders: {'Status': '200', 'Date': 'now(0)', 'Expires': 'now(100)' } },
 { responseHeaders: {'Status': '304', 'Date': 'now(0)', 'Expires': 'now(100)' }, return200AtFirstResponse: true },
 { responseHeaders: {'Status': '200', 'Date': 'now(0)', 'Expires': 'now(-100)' } },
 { responseHeaders: {'Status': '304', 'Date': 'now(0)', 'Expires': 'now(-100)' }, return200AtFirstResponse: true },
];

description("Test of conditional request.");

debug("running " + tests.length + " tests");
debug("");

runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
