5.53/LayoutTests/http/tests/cache/disk-cache/manx/disk-cache-304-status-code-expected.txt

Test of conditional request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 4 tests

response headers: {"Status":"200","Date":"now(0)","Expires":"now(100)"}
response source: Disk cache

response headers: {"Status":"304","Date":"now(0)","Expires":"now(100)"}
Status code of the first response is overriden by 200
response source: Disk cache

response headers: {"Status":"200","Date":"now(0)","Expires":"now(-100)"}
response source: Network

response headers: {"Status":"304","Date":"now(0)","Expires":"now(-100)"}
Status code of the first response is overriden by 200
response source: Disk cache after validation

PASS successfullyParsed is true

TEST COMPLETE
